+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Display warning message when user tries to add a field without soil data as well as in the summary page for such a field. [CCROP-317](https://opensource.ncsa.illinois.edu/jira/browse/CCROP-317)
 - Validation on cover crop and cash crop planting dates when running a job [CCROP-316](https://opensource.ncsa.illinois.edu/jira/browse/CCROP-316)
+- "Growing Degree Days" popup graph to the Dashboard [CCROP-319](https://opensource.ncsa.illinois.edu/jira/browse/CCROP-319)
 
 ## [1.0.0] - 2020-09-30